[MachineLICM] Improve register pressure estimation.
Sometimes a definition in loop's preheader is only used outside of the
loop. If we run out of registers inside the loop, such definitions are
going to be spilled first, so they should not count when we estimate the
register pressure inside the loop. Note that live-through registers are
not counted as well.